+++ dart/test/mocks.dart
@@ -8,6 +8,8 @@ class MockTransport extends Mock implements Transport {}
 
 class MockHub extends Mock implements Hub {}
 
+class MockIntegration extends Mock implements Integration {}
+
 final fakeDsn = 'https://abc@def.ingest.sentry.io/1234567';
 
 final fakeException = Exception('Error');