+++ flutter/lib/src/sentry_flutter.dart
@@ -40,7 +40,7 @@ mixin SentryFlutter {
       appRunner,
       // first integration to add
       // will catch any errors that may occur in the Flutter framework itself.
-      [flutterErrorIntegration],
+      [FlutterErrorIntegration()],
     );
   }
 
@@ -116,17 +116,16 @@ mixin SentryFlutter {
     // that allow us to send events to the network and then the Flutter integrations.
     // Flutter Web doesn't need that, only Android and iOS.
     if (!kIsWeb) {
-      options.addIntegration(nativeSdkIntegration(options, _channel));
+      options.addIntegration(NativeSdkIntegration(_channel));
     }
 
     // will enrich the events with the device context and native packages and integrations
     if (isIOS()) {
-      options.addIntegration(loadContextsIntegration(options, _channel));
+      options.addIntegration(LoadContextsIntegration(_channel));
     }
 
     if (isAndroidChecker()) {
-      options
-          .addIntegration(loadAndroidImageListIntegration(options, _channel));
+      options.addIntegration(LoadAndroidImageListIntegration(_channel));
     }
   }